Extract platform plugins deployment
