+++ src/deployers/basic_plugin_deployer.h
@@ -3,33 +3,43 @@
 // system headers
 #include <memory>
 
+// library headers
+#include <linuxdeploy/core/appdir.h>
+
 // local headers
 #include "plugin_deployer.h"
 
-using namespace linuxdeploy::plugin::qt;
 
-/**
- *
- */
-class BasicPluginDeployer : public PluginDeployer {
-private:
-    class Private;
-    std::shared_ptr<Private> d;
+namespace linuxdeploy {
+    namespace plugin {
+        namespace qt {
+            /**
+             *
+             */
+            class BasicPluginDeployer : public PluginDeployer {
+            protected:
+                std::string moduleName;
+                core::appdir::AppDir& appDir;
+                boost::filesystem::path qtPluginsPath;
 
-public:
-    /**
-     * Default constructor. Constructs a basic deployer for a plugin with the given name.
-     *
-     * @param pluginName
-     */
-    explicit BasicPluginDeployer(const std::string& pluginName);
+            public:
+                /**
+                 * Default constructor. Constructs a basic deployer for a plugin with the given name.
+                 *
+                 * @param moduleName
+                 */
+                explicit BasicPluginDeployer(std::string moduleName, core::appdir::AppDir& appDir,
+                                             boost::filesystem::path qtPluginsPath);
 
-    /**
-     * Default destroyer is good enough for this class for now, but in case we need to change this we declare a virtual
-     * one.
-     */
-    virtual ~BasicPluginDeployer() = default;
+                /**
+                 * Default destroyer is good enough for this class for now, but in case we need to change this we declare a virtual
+                 * one.
+                 */
+                virtual ~BasicPluginDeployer() = default;
 
-public:
-    void deploy() override;
-};
+            public:
+                bool deploy() override;
+            };
+        }
+    }
+}
