+++ src/deployers/plugin_deployer_factory.h
@@ -4,6 +4,10 @@
 #include <memory>
 #include <string>
 
+// library headers
+#include <linuxdeploy/core/appdir.h>
+#include <boost/filesystem.hpp>
+
 // local headers
 #include "plugin_deployer.h"
 
@@ -12,7 +16,9 @@ namespace linuxdeploy {
         namespace qt {
             class PluginDeployerFactory {
             public:
-                static std::shared_ptr<PluginDeployer> getInstance(const std::string& moduleName);
+                static std::shared_ptr<PluginDeployer>
+                getInstance(const std::string& moduleName, core::appdir::AppDir& appDir,
+                            const boost::filesystem::path& qtPluginsPath);
             };
         }
     }
